pyproject.toml
[project]
name = "backend-improving"
version = "0.1.0"
description = "Backend improvement project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery[redis,sqs]>=5.5.2",
    "coverage>=7.9.1",
    "django>=5.2",
    "django-celery-beat>=2.8.0",
    "django-debug-toolbar>=5.2.0",
    "django-filter>=25.1",
    "django-split-settings>=1.3.2",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "pillow>=11.2.1",
    "psycopg[binary]>=3.2.6",
    "pytest-cov>=6.2.1",
    "python-decouple>=3.8",
    "requests>=2.32.3",
    "ruff>=0.11.7",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
    "ruff>=0.11.7",
]

[tool.ruff]
line-length = 119

exclude = [
    ".venv",
    "venv",
    "__pycache__",
    "migrations",
    "settings.py",
]

[tool.black]
line-length = 119

[tool.flake8]
max-line-length = 119

[tool.ruff.format]

[tool.ruff.lint]
select = ["E", "F", "I", "D"]  # "I" enables import sorting
ignore = ["D203", "D209", "D210", "D212", "D213", "D200", "D413", "D104", "D401"]  # Ignore docstring formatting rules
